Move MultiError to types package
