+++ utils.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"fmt"
-	"sync"
 	"unicode"
 )
 
@@ -128,37 +127,4 @@ func removeInvalidTargets(targets []string) []string {
 	}
 
 	return filteredTargets
-}
-
-type MultiError struct {
-	Errors []error
-	mux    sync.Mutex
-}
-
-func (err *MultiError) Error() string {
-	str := ""
-
-	for _, e := range err.Errors {
-		str += e.Error() + "\n"
-	}
-
-	return str[:len(str)-1]
-}
-
-func (err *MultiError) Add(e error) {
-	if e == nil {
-		return
-	}
-
-	err.mux.Lock()
-	err.Errors = append(err.Errors, e)
-	err.mux.Unlock()
-}
-
-func (err *MultiError) Return() error {
-	if len(err.Errors) > 0 {
-		return err
-	}
-
-	return nil
 }
